refactor: moved each tauri-plugin to own rust-crate
