+++ src-tauri/cabr2_config/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "cabr2_config"
+version = "0.0.0"
+authors = [
+  "crapStone <crapstone01@gmail.com>",
+  "6543 <6543@obermui.de>",
+]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base64 = "0.13.0"
+cabr2_types = { path = "../cabr2_types/" }
+directories-next = "2.0.0"
+lazy_static = "1.4.0"
+log = "0.4.14"
+serde = { version = "1.0.123", features = ["derive"] }
+serde_json = "1.0.64"
+tauri = { version = "0.11.1", default-features = false }
+thiserror = "1.0.24"
+toml = { version = "0.5.8", features = ["preserve_order"] }
+
+[features]
+portable = [ ]
