+++ src-tauri/cabr2_load_save/src/lib.rs
@@ -1,3 +1,6 @@
+#![allow(clippy::new_without_default)]
+#![allow(clippy::unnecessary_unwrap)]
+
 use tauri::plugin::Plugin;
 
 mod cmd;
