+++ src-tauri/cabr2_logger/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "cabr2_logger"
+version = "0.0.0"
+authors = [
+  "crapStone <crapstone01@gmail.com>",
+  "6543 <6543@obermui.de>",
+]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cabr2_config = { path = "../cabr2_config/" }
+cabr2_types = { path = "../cabr2_types/" }
+chrono = "0.4.19"
+fern = "0.6.0"
+log = "0.4.14"
+serde = { version = "1.0.123", features = ["derive"] }
+serde_json = "1.0.64"
+tauri = { version = "0.11.1", default-features = false }
