+++ src-tauri/cabr2_logger/src/cmd.rs
@@ -1,14 +1,7 @@
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::Value;
 
-#[derive(Debug, Deserialize, Serialize)]
-pub enum LogLevel {
-  TRACE,
-  DEBUG,
-  INFO,
-  WARNING,
-  ERROR,
-}
+use cabr2_types::logging::LogLevel;
 
 #[derive(Deserialize)]
 #[serde(tag = "cmd", rename_all = "camelCase")]
