+++ src-tauri/cabr2_search/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "cabr2_search"
+version = "0.0.0"
+authors = [
+  "crapStone <crapstone01@gmail.com>",
+  "6543 <6543@obermui.de>",
+]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cabr2_types = { path = "../cabr2_types/" }
+chrono = "0.4.19"
+lazy_static = "1.4.0"
+log = "0.4.14"
+roxmltree = "0.14.0"
+serde = { version = "1.0.123", features = ["derive"] }
+serde_json = "1.0.64"
+tauri = { version = "0.11.1", default-features = false }
+thiserror = "1.0.24"
+ureq = { version = "1.5.4", features = ["tls", "json"] }
