+++ src-tauri/cabr2_search/src/gestis/error.rs
