+++ src-tauri/cabr2_search/src/lib.rs
@@ -1,0 +1,6 @@
+#![allow(clippy::new_without_default)]
+#![allow(clippy::unnecessary_unwrap)]
+
+mod gestis;
+
+pub use gestis::Gestis;
