+++ src-tauri/cabr2_types/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cabr2_types"
+version = "0.0.0"
+authors = [
+  "crapStone <crapstone01@gmail.com>",
+  "6543 <6543@obermui.de>",
+]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = { version = "0.4.19", features = ["serde"] }
+log = "0.4.14"
+serde = { version = "1.0.123", features = ["derive"] }
