+++ src-tauri/src/main.rs
@@ -1,16 +1,12 @@
 #![cfg_attr(all(not(debug_assertions), target_os = "windows"), windows_subsystem = "windows")]
 
-mod config;
-mod load_save;
-mod logger;
-mod search;
-mod types;
-
 fn main() {
-  let logger = logger::Logger::new();
-  let config = config::Config;
-  let load_save = load_save::LoadSave::new();
-  let search = search::Gestis::new();
+  // must be initialized first
+  let logger = cabr2_logger::Logger::new();
+
+  let config = cabr2_config::Config;
+  let load_save = cabr2_load_save::LoadSave::new();
+  let search = cabr2_search::Gestis::new();
 
   log::debug!("initializing tauri application...");
