--- src-tauri/src/search/mod.rs
@@ -1,3 +1,0 @@
-mod gestis;
-
-pub use gestis::Gestis;
